requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests>=2.32.4
httpx>=0.24.0
python-dotenv>=1.1.1
pydantic>=2.11.7
pydantic_core==2.33.2
fastapi-jsonrpc==3.4.1
grpcio==1.74.0
grpcio-status==1.74.0
jsonrpclib-pelix==0.4.3.4
ipython-sql==0.5.0
sqlalchemy-migrate==0.11.0
sqlparse==0.5.3